// apps/server/src/generated/prisma/enums.ts
export const AccountType = {
  ASSET: 'ASSET',
  LIABILITY: 'LIABILITY',
  EQUITY: 'EQUITY',
  REVENUE: 'REVENUE',
  EXPENSE: 'EXPENSE'
} as const

export type AccountType = (typeof AccountType)[keyof typeof AccountType]
// ...
export const JournalSource = {
  MANUAL: 'MANUAL',
  SALE: 'SALE',
  PURCHASE: 'PURCHASE',
  PAYMENT: 'PAYMENT',
  EXPENSE: 'EXPENSE',
  RETURN: 'RETURN',
  REFUND: 'REFUND',
  OPENING: 'OPENING',
  CLOSING: 'CLOSING',
  STOCKTAKE: 'STOCKTAKE'
} as const

export type JournalSource = (typeof JournalSource)[keyof typeof JournalSource]